New: user register, database connect

+++ auth0rize.auth.domain/User/User.cs
@@ -11,7 +11,7 @@
         public byte[] Salt { get; set; } = null!;
         public bool IsDoubleFactorActivate { get; set; } = false;
         public string Avatar { get; set; } = null!;
-        public int Type { get; set; }
+        public long Type { get; set; }
         public long Domain {  get; set; }
     }
 }
